Add confirm delete conversation dialog

+++ src/elements/Toolbar.tsx
@@ -1,4 +1,5 @@
 import { PanelLeft, Trash2, MailPlus } from "lucide-react";
+import { confirm } from "@tauri-apps/plugin-dialog";
 import { useConversationStore } from "../store/conversationStore";
 import { useShallow } from "zustand/react/shallow";
 
@@ -10,15 +11,24 @@ const Toolbar = ({ toggleSidebar }: ToolbarProps) => {
   const [addConversation, deleteActiveConversation] = useConversationStore(
     useShallow((state) => [
       state.addConversation,
-      state.deleteActiveConversation
+      state.deleteActiveConversation,
     ])
   );
 
+  const removeConversation = async () => {
+    console.log("removing...");
 
-  const removeConversation = () => {
-    deleteActiveConversation();
-  }
+    // Creates a confirmation Ok/Cancel dialog
+    const confirmation = await confirm(
+      "Are you sure you want to delete this conversation?",
+      { title: "Delete Conversation?", kind: "warning" }
+    );
 
+    // Prints boolean to the console
+    if (confirmation) {
+      deleteActiveConversation();
+    }
+  };
 
   return (
     <div className="flex justify-between items-center p-2 bg-gray-100 text-blue-500">
